Project2: Lab13: Do 'Tema 1'
This implements user roles by using the `uuid` library: for each login,
a UUID is generated and stored in a map that has the UUID as the key and
the user role as the value. The user session object is extended to store
the generated UUID as well.

This should be safe, as the UUID is a v4 UUID which is truly random, so
it cannot be predicted or guessed by adversarial users.